Proposed change to event view screen
This enables click anywhere to close rather than the single x in the
corner.  This is much easier for a mobile layout.  Just a thought.  The
original code is still there, just commented out.  I will help with the
database integration for you.  Just to clarify... IS THE "joinScreen"
folder not being used?