Various style changes and addition to Jessicas code
Changed the style of the header bar to make text always centered minus
the space of the menu button.  Removed "menu" text from the menu button
to conserve space on mobile.

Additions to Jessica's code include:
Made menu's close after choosing a selection, and moved function to
change intervals out of main function for counting to reduce redundancy
in the AJAX call to the server once implemented.